Parameterise "Use TCP" test as as precursor to and POC for IPv4 test
Update tests[13] to use function pointers and other parameters, so no need
to check tests[13] where we check for tests[6, 9 &c].

There was "mystery code" that re-fetched pointers after test 12
so removed that as well (IIRC that may hark back to the days of doing a delayed
copy on a mangle that increase packet size).

Tests 18 & 11 sometimes fail if you blast lots of e.g. ASDASD lines at it.
Think this is an old bug.